task2/requirements.txt
Django==3.2.15
python-dateutil==2.8.1
psycopg2-binary>=2.8